requirements.txt
webdriver-manager==1.7
eyes-selenium==4.2.1
eyes-common==4.2.1
eyes-core==4.2.1
requests
pytest
attrs
py
cffi
pytest-ordering
pytest-xdist
beautifulsoup4
pytest-rerunfailures
pytz
virtualenv
pytest-progress
pytest-html
PyYAML
selenium==3.141.0
pdbpp
requests-toolbelt
delayed_assert
softest